Backend/requirements.txt
pymongo==4.6.1
Flask
python-dotenv
pymongo
Werkzeug
Flask-JWT-Extended